refactor(scales): move things around and work with other packages
