+++ packages/scales/src/bandScale.ts
@@ -1,12 +1,16 @@
 import { scaleBand } from 'd3-scale'
-import { ComputedSerieAxis, ScaleBand, ScaleBandSpec, StringValue } from './types'
+import { ComputedSerieAxis, ScaleBand, ScaleBandSpec, StringValue, ScaleAxis } from './types'
 
 export const createBandScale = <Input extends StringValue>(
-    _spec: ScaleBandSpec,
+    { round = true }: ScaleBandSpec,
     data: ComputedSerieAxis<Input>,
-    size: number
+    size: number,
+    axis: ScaleAxis
 ) => {
-    const scale = scaleBand<Input>().range([0, size]).domain(data.all)
+    const scale = scaleBand<Input>()
+        .range(axis === 'x' ? [0, size] : [size, 0])
+        .domain(data.all)
+        .round(round)
 
     const typedScale = scale as ScaleBand<Input>
     typedScale.type = 'band'
