+++ packages/scales/src/compute.ts
@@ -4,8 +4,13 @@ import sortBy from 'lodash/sortBy'
 import last from 'lodash/last'
 import isDate from 'lodash/isDate'
 import { createDateNormalizer } from './timeHelpers'
-import { ScaleAxis, ScaleSpec, Series, ScaleValue, SerieAxis } from './types'
-import { computeScale } from './computeScale'
+import { ScaleAxis, ScaleSpec, Series, ScaleValue, SerieAxis, ComputedSerieAxis } from './types'
+import { createLinearScale } from './linearScale'
+import { createPointScale } from './pointScale'
+import { createBandScale } from './bandScale'
+import { createTimeScale } from './timeScale'
+import { createLogScale } from './logScale'
+import { createSymLogScale } from './symLogScale'
 
 type XY = ReturnType<typeof generateSeriesXY>
 
@@ -48,6 +53,30 @@ export const getOtherAxis = (axis: ScaleAxis): ScaleAxis => (axis === 'x' ? 'y' 
 
 export const compareValues = (a: string | number, b: string | number) => a === b
 export const compareDateValues = (a: Date, b: Date) => a.getTime() === b.getTime()
+
+export function computeScale<Input extends ScaleValue>(
+    spec: ScaleSpec,
+    data: ComputedSerieAxis<any>,
+    size: number,
+    axis: ScaleAxis
+) {
+    switch (spec.type) {
+        case 'linear':
+            return createLinearScale(spec, data, size, axis)
+        case 'point':
+            return createPointScale<Input>(spec, data, size)
+        case 'band':
+            return createBandScale<Input>(spec, data, size, axis)
+        case 'time':
+            return createTimeScale(spec, data, size)
+        case 'log':
+            return createLogScale(spec, data, size, axis)
+        case 'symlog':
+            return createSymLogScale(spec, data, size, axis)
+        default:
+            throw new Error('invalid scale spec')
+    }
+}
 
 export const computeXYScalesForSeries = (
     _series: XYSeries[],
